+++ sdk/securityinsight/arm-securityinsight/src/operationsInterfaces/index.ts
@@ -37,6 +37,8 @@ export * from "./threatIntelligenceIndicators";
 export * from "./threatIntelligenceIndicatorMetrics";
 export * from "./watchlists";
 export * from "./watchlistItems";
+export * from "./workspaceManagerAssignments";
+export * from "./workspaceManagerAssignmentJobs";
 export * from "./dataConnectors";
 export * from "./dataConnectorsCheckRequirementsOperations";
 export * from "./operations";